feat: уменьшил количество попыток на отгадку

+++ lw1/hangman/src/viewModel/HangmanViewModel.hpp
@@ -99,7 +99,7 @@ public:
 		return states;
 	}
 
-	int GetWrongGuesses() const
+	size_t GetWrongGuesses() const
 	{
 		return m_model.GetWrongGuesses();
 	}
